Force the logs to be written in UTF-8 and not using platform encoding
Caveat: to test this, one *must* set -Dfile.encoding=ISO-8859-1 or anything that is not UTF-8.
Then run the test.

On my machine, it was failing before I add the `fileHandler.setEncoding("UTF-8")` call (see next commit).

This commit is *expected to fail* since it only contains a refactoring and a test, not the fix yet.